fikset masse.. Har laget kalkulasjon av antall timer arbeidet, har lagt til beskrivelse prosjekt, har endret på visning av timeliste

+++ app/timetracking/models.py
@@ -22,8 +22,7 @@ class Timetracking(PersistentModel):
     time_end = models.CharField(max_length=5)
     description = models.TextField()
     
-    hours_worked = models.IntegerField()
-    
+    hours_worked = models.DecimalField(decimal_places=3, max_digits=5)
     
     def __unicode__(self):
         return unicode(self.date)
